add Table Member to common component and remove it from new and edit project Relates #86
