--- client/src/components/Projects/ProjectEdit/TableMember/logic.js
@@ -1,9 +1,0 @@
-const filterIdMember = teamMember => {
-  const filter = [];
-  teamMember.forEach(obj => {
-    filter.push(obj.id.toString());
-  });
-  return filter;
-};
-
-export default filterIdMember;
