--- client/src/components/Projects/ProjectEdit/TableMember/projects.json
@@ -1,19 +1,0 @@
-[
-  {
-    "id": "1",
-    "name": "UN project",
-    "description": "project",
-    "teamMember": [
-      {
-        "id": "1",
-        "name": "amin",
-        "role": "developer"
-      },
-      {
-        "id": "2",
-        "name": "mohammed",
-        "role": "sucrm"
-      }
-    ]
-  }
-]
