--- client/src/components/Projects/ProjectNew/TableMember/style.css
@@ -1,15 +1,0 @@
-.table-font-row {
-  font-size: 1.2rem;
-  font-weight: 400;
-}
-
-.table-font-header th {
-  font-size: 1.2rem;
-  background: rgb(228, 225, 225) !important;
-  border-radius: 0 !important;
-}
-
-.ant-table-fixed-header .ant-table-scroll .ant-table-header {
-  overflow: hidden !important;
-  margin-bottom: 0px !important;
-}
